backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.28.0
python-dotenv==1.0.1
apify-client==1.8.1
pydantic==2.10.3
